@@ -48,7 +48,8 @@ public class HorizontalVideoAdapter extends RecyclerView.Adapter<HorizontalVideo
     public void onBindViewHolder(@NonNull horzonVideoVh holder, int position) {
         Video video = videoList.get(position);
 
-//        holder.txt_video_name.setText(video.getVideoTitle());
+
+
         Picasso.get().load(video.getVideoThumbnailB()).into(holder.img_video);
 
         holder.rel_video.setOnClickListener(new View.OnClickListener() {
@@ -74,13 +75,11 @@ public class HorizontalVideoAdapter extends RecyclerView.Adapter<HorizontalVideo
 
     class horzonVideoVh extends RecyclerView.ViewHolder{
         AppCompatImageView img_video;
-//        AppCompatTextView txt_video_name;
         RelativeLayout rel_video;
         public horzonVideoVh(@NonNull View itemView) {
             super(itemView);
 
             img_video = itemView.findViewById(R.id.img_video);
-           // txt_video_name = itemView.findViewById(R.id.txt_video_name);
             rel_video = itemView.findViewById(R.id.rel_video);
         }
     }
